fix: model example
